backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.25
pymysql==1.0.2
python-decouple==3.4
